requirements.txt
# Core ML and Computer Vision
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Web Frameworks
streamlit>=1.28.0
gradio>=4.0.0

# Utilities
pyyaml>=6.0
requests>=2.31.0

